+++ app/src/main/java/org/breezyweather/Migrations.kt
@@ -20,6 +20,7 @@ import android.content.Context
 import org.breezyweather.background.forecast.TodayForecastNotificationJob
 import org.breezyweather.background.forecast.TomorrowForecastNotificationJob
 import org.breezyweather.background.weather.WeatherUpdateJob
+import org.breezyweather.db.repositories.AlertEntityRepository
 import org.breezyweather.db.repositories.LocationEntityRepository
 import org.breezyweather.db.repositories.WeatherEntityRepository
 import org.breezyweather.settings.SettingsManager
@@ -44,6 +45,10 @@ object Migrations {
                     // - old current location weather data that was kept
                     LocationEntityRepository.regenerateAllFormattedId()
                     WeatherEntityRepository.deleteAllWeather()
+                }
+                if (oldVersion < 40610) {
+                    // Clean up all alerts due to change of alertId from long to string
+                    AlertEntityRepository.deleteAllAlerts()
                 }
             }